LIME Implementation and general improvements
LIME has now been implemented in model_evaluation and appears to be fully functional. It can probably use some tweaks in terms of parameters and how it is displayed, but the base is there.

Additionally, some general improvements/changes to the code of both the model_evaluation and model_training notebooks were made and are being committed.